Fix ProxyActivity port typing and a decode() default-clobbering bug
ProxyActivity.port/proxy_port/listen_port decoded to floats (8080.0)
because they were typed float instead of int, unlike every other port
field in the SDK.

decode() also passed every missing wire field as an explicit None
keyword argument, even fields with a dataclass default. That's
harmless today since no model field declares one, but it would
silently override any default added in the future. Missing fields are
now omitted from the constructor call when the field has a default (or
default_factory), so the model's own default applies.

+++ src/webshare/types/proxy.py
@@ -235,10 +235,10 @@ class ProxyActivity:
     ip_address: str | None
     hostname: str | None
     domain: str | None
-    port: float | None
-    proxy_port: float | None
+    port: int | None
+    proxy_port: int | None
     listen_address: str | None
-    listen_port: float | None
+    listen_port: int | None
 
 
 @dataclass
